+++ stubs/main.ts
@@ -1,3 +1,4 @@
-import { getDirname } from '@poppinss/utils'
+import { dirname } from 'node:path'
+import { fileURLToPath } from 'node:url'
 
-export const stubsRoot = getDirname(import.meta.url)
+export const stubsRoot = dirname(fileURLToPath(import.meta.url))